# API/userAPI/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "User API Documentation",
    "description": "Documentation de l'API User",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:3003/api/users/v1"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/tryLogin": {
      "post": {
        "summary": "Authentification de l'utilisateur",
        "description": "Authentifie l'utilisateur avec les informations d'identification fournies.",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "Adresse e-mail de l'utilisateur."
                  },
                  "password": {
                    "type": "string",
                    "description": "Mot de passe de l'utilisateur."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentification réussie.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message de succès."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Paramètres manquants dans le corps de la requête.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message d'erreur."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Compte inexistant.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message d'erreur."
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Mot de passe incorrect.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message d'erreur."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/create/user": {
      "post": {
        "summary": "Création de compte utilisateur",
        "description": "Crée un nouveau compte utilisateur avec les informations fournies.",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "Adresse e-mail du nouvel utilisateur."
                  },
                  "password": {
                    "type": "string",
                    "description": "Mot de passe du nouvel utilisateur."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Compte créé avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message de succès."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Paramètres manquants dans le corps de la requête.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message d'erreur."
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Compte déjà existant.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message d'erreur."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/update/password": {
      "put": {
        "summary": "Modification du mot de passe de l'utilisateur",
        "description": "Modifie le mot de passe de l'utilisateur avec le nouveau mot de passe fourni.",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "Adresse e-mail de l'utilisateur."
                  },
                  "password": {
                    "type": "string",
                    "description": "Mot de passe actuel de l'utilisateur."
                  },
                  "new_password": {
                    "type": "string",
                    "description": "Nouveau mot de passe de l'utilisateur."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mot de passe changé avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message de succès."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Paramètres manquants dans le corps de la requête.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message d'erreur."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Utilisateur inexistant.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message d'erreur."
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Mot de passe incorrect.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message d'erreur."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/delete/user": {
      "delete": {
        "summary": "Suppression de l'utilisateur",
        "description": "Supprime l'utilisateur avec l'adresse e-mail fournie.",
        "tags": [
          "User"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "Adresse e-mail de l'utilisateur à supprimer."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Utilisateur supprimé avec succès.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message de succès."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Paramètres manquants dans le corps de la requête.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message d'erreur."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}